moved relief attempts component into loc directory
